+++ Sources/Utility/Constants.swift
@@ -9,7 +9,7 @@ public var log = Logger(outputType: .console)
 /// Constants to reference across the project.
 public enum Constants {
     /// The current tool version string. Conforms to SemVer 2.0.
-    public static let currentVersion: String = "0.5.0"
+    public static let currentVersion: String = "0.6.0"
 
     /// The name of this tool.
     public static let toolName: String = "AnyLint"
@@ -22,4 +22,13 @@ public enum Constants {
 
     /// The validate-only mode argument for command-line pass-through.
     public static let validateArgument: String = "validate"
+
+    /// The separator indicating that next come regex options.
+    public static let regexOptionsSeparator: String = #"\"#
+
+    /// Hint that the case insensitive option should be active on a Regex.
+    public static let caseInsensitiveRegexOption: String = "i"
+
+    /// Hint that the case dot matches newline option should be active on a Regex.
+    public static let dotMatchesNewlinesRegexOption: String = "m"
 }